:recycle: api: move CompetenceMark model from results to shared
