+++ api/src/certification/course/infrastructure/repositories/course-assessment-result-repository.js
@@ -4,7 +4,7 @@ import { knex } from '../../../../../db/knex-database-connection.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import { NotFoundError } from '../../../../shared/domain/errors.js';
 import { AssessmentResult } from '../../../../shared/domain/models/AssessmentResult.js';
-import { CompetenceMark } from '../../../results/domain/models/CompetenceMark.js';
+import { CompetenceMark } from '../../../shared/domain/models/CompetenceMark.js';
 import { JuryComment, JuryCommentContexts } from '../../../shared/domain/models/JuryComment.js';
 
 function _toDomain({ assessmentResultDTO, competencesMarksDTO }) {
