+++ api/tests/certification/results/unit/domain/models/CompetenceMark_test.js
@@ -1,5 +1,5 @@
 import { ObjectValidationError } from '../../../../../../lib/domain/errors.js';
-import { CompetenceMark } from '../../../../../../src/certification/results/domain/models/CompetenceMark.js';
+import { CompetenceMark } from '../../../../../../src/certification/shared/domain/models/CompetenceMark.js';
 import { catchErr, domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | Competence Mark', function () {
